test(reader): split navigation bridge payload coverage
Why:
- Reader navigation bridge and document-payload contracts are BibleUI package behavior, not app-host bootstrap behavior.
- Keeping these tests in the app-host bundle preserves the slow simulator app-build path the migration is removing.

What Changed:
- Moved compare bridge payload, reader document JSON factory, auxiliary fallback document, and rendered-content token tests into BibleUITests.
- Left AndBibleTests+ReaderNavigation focused on the remaining app-host migration slices.
- Updated the test architecture migration plan with the completed reader-navigation bridge slice.

Validation:
- GitNexus impact checks for moved symbols reported LOW risk with no affected execution flows.
- GitNexus staged change detection reported no mapped production-flow changes.
- Ran focused BibleUITests ReaderNavigationBridgePayloadTests package suite.
- Ran focused AndBibleTests reader-navigation remainder compile sentinel.
- Ran git diff checks, docblock guard, source-guards, and adversarial subagent review.

Impact:
- Reduces app-host unit-test surface while preserving reader bridge contract coverage in the package test lane.
- No production behavior changes.
